Initial commit containing data and readme
Often the very first commit message is something
similar to "Initial commit". In general, we want
to make our commit messages informative.

The first line is important to keep short (<50 chars)
and to the point. This is a line that will appear
in your logs, and should be enough to determine if
this is the commit that you want to look at.

The message below the body can be expanded to be
detailed. In this case, the commit message is more
a "meta-commit" messages, telling you about commit
messages in general. Some general points about
commit messages:

- The first line should be a short summary (<50 chars)
- The first line should be written in imperitive form.
- Include a blank line after the first line summary
  before going onto your more detailed commits.
- Wrap all text at 72 chars (to make scanning the log
  easy, even on older terminals)