Add Excel regression analysis, adj R2 = 0.733
MeticsL
- R2: 0.741
- Adj R2: 0.733
- Std Err: 4.745
- Nobs: 506

Included a simple linear regression analysis in an Excel workbook.

There is no train/test split, but the traditional Excel analysis
includes R squared and adjusted R squared values.

If using multiple models, or iterations, pick a standard way of
reporting your different metrics. The idea is to make the logs
easily "scannable" so that you can review logs and quickly compare
models. Sometimes it can be hard, as some metrics only make sense
in some settings. For example, if we move to Machine Learning models,
we would eliminate adjusted R2 (but we would probably replace it with
validation R2)