Replicate Excel analysis in Python
R^2: 0.740642
n_obs: 507

We have duplicated our analysis from Excel in Python. We still
have not performed a train-test split, and we have not calcuated
the adjusted R^2, so we don't have a good feeling foe whether or
not this model is overfitting.

Note that the commit messages don't include a list of files added
or a list of files changed. Git keeps track of all of that for us.
Instead, in a commit message we should focus on what we did, not
how we did it.

If we need to rollback changes, Git can take care of that for us