Change model from Linear Model to LassoCV
- R^2 train: 0.7145
- R^2 test:  0.6642
- Observations: 507

Introduced a train-test split, and used LassoCV to regularize our
result. The model performed worse than the linear regression model
(R^2 = 0.7406). It is hard to know how reliable this R^2 value
from Linear Regression is, as it is the R^2 on the training data.

We see that the R^2 in LassoCV is worse for the training data than
the R^2 for Linear Regression. We should expect this: regularization
makes performance worse on the training set in the hopes that the
test set R^2 will be better (i.e. we are aiming to improve the
generalizability of the model)

It is a little more concerning that our test set R^2 (0.6642) does
so poorly in comparison to the adjusted R^2 of the Linear Regression
model.