Refactor bootstrap to separate file
Moved the bootstrapping code to a separate file. We also looked at
the coefficents in "natural" units (i.e. the original units of the
feature) as we felt that for reporting purposes, these would be more
easily interpretable.

For determining feature importance, we still use the coefficients of
the scaled features.