Add regression on log(price)
The prices are strongly right skewed, so we tried applying a log
and fitting to the log of price instead. The metric was still the
MSE or R^2 value on the original (non-logged) prices.

We found this model did slightly worse than (i.e. < 1 percentage
point difference) than working directly on the non-logged prices.

Recommend no change.